// src/core/api/generated/spring/schemas/OrderAddressDetailDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export interface OrderAddressDetailDTO {
  id?: number;
  orderId?: number;
  /**
   * @minLength 0
   * @maxLength 100
   */
  shipToFirstName?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  shipToMiddleName?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  shipToLastName?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  shipToAddLine1?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  shipToAddLine2?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  shipToCity?: string;
  /**
   * @minLength 0
   * @maxLength 10
   */
  shipToZipcode?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  shipToContact?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  shipToState?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  shipToCountry?: string;
  billToSameFlag?: boolean;
  /**
   * @minLength 0
   * @maxLength 100
   */
  billToFirstName?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  billToMiddleName?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  billToLastName?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  billToAddLine1?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  billToAddLine2?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  billToCity?: string;
  /**
   * @minLength 0
   * @maxLength 10
   */
  billToZipcode?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  billToContact?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  billToState?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  billToCountry?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  updatedBy?: string;
  lastUpdated?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  createdBy?: string;
  createdDate?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  lastModifiedBy?: string;
  lastModifiedDate?: string;
}