// src/core/api/generated/spring/schemas/OrderHistoryDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export interface OrderHistoryDTO {
  id?: number;
  orderId?: number;
  /**
   * @minLength 0
   * @maxLength 50
   */
  status?: string;
  notificationSent?: boolean;
  /**
   * @minLength 0
   * @maxLength 50
   */
  updatedBy?: string;
  lastUpdated?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  createdBy?: string;
  createdDate?: string;
  /**
   * @minLength 0
   * @maxLength 50
   */
  lastModifiedBy?: string;
  lastModifiedDate?: string;
}